requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.3.0
numpy>=1.21.0
pydantic>=1.8.0
requests>=2.26.0
python-dotenv>=0.19.0

# YOLO and AI dependencies
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0

# RAG Agent dependencies
beautifulsoup4>=4.12.0
google-search-results>=2.4.2
openai>=0.27.0,<1.0.0
anthropic>=0.3.0
transformers>=4.21.0

# Optional dependencies for local models
# ollama (install separately if using local models)
